refactor: move FDM grid routines to new class
refactor: include all ellipsoidal properties in grid class

refactor: changed variable for setting global grid flag to is_global
