+++ doc/source/api_reference/solve/grid.rst
@@ -1,0 +1,15 @@
+====
+grid
+====
+
+Class for setting up finite difference grids for tidal modeling
+
+`Source code`__
+
+.. __: https://github.com/tsutterley/pyTMD/blob/main/pyTMD/solve/grid.py
+
+General Attributes and Methods
+==============================
+
+.. autoclass:: pyTMD.solve.grid
+   :members:
