+++ doc/source/api_reference/solve/solve.rst
@@ -6,4 +6,5 @@ solve
     :maxdepth: 1
 
     constants.rst
+    grid.rst
     model.rst
