+++ pyTMD/solve/__init__.py
@@ -1,2 +1,3 @@
 from .constants import *
+from .grid import grid
 from .model import model
